refactor:PaymentService 를 Client 로 변경
