+++ src/main/java/roomescape/payment/client/PaymentClient.java
@@ -1,0 +1,39 @@
+package roomescape.payment.client;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.nio.charset.StandardCharsets;
+
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestClient;
+import roomescape.common.exception.PaymentException;
+import roomescape.reservation.dto.request.TossPaymentConfirmRequest;
+import roomescape.reservation.dto.response.TossErrorResponse;
+import roomescape.reservation.dto.response.TossPaymentResponse;
+
+@Component
+public class PaymentClient {
+
+    private final RestClient tossRestClient;
+    private final ObjectMapper objectMapper;
+
+    public PaymentClient(RestClient tossRestClient, ObjectMapper objectMapper) {
+        this.tossRestClient = tossRestClient;
+        this.objectMapper = objectMapper;
+    }
+
+    public TossPaymentResponse confirmPayment(TossPaymentConfirmRequest request) {
+        return tossRestClient.post()
+                .uri("/payments/confirm")
+                .body(request)
+                .retrieve()
+                .onStatus(
+                        status -> status.is4xxClientError() || status.is5xxServerError(),
+                        (req, res) -> {
+                            String errorBody = new String(res.getBody().readAllBytes(), StandardCharsets.UTF_8);
+                            TossErrorResponse errorResponse = objectMapper.readValue(errorBody, TossErrorResponse.class);
+                            throw new PaymentException("결제 실패 : " + errorResponse.message());
+                        }
+                )
+                .body(TossPaymentResponse.class);
+    }
+}
