--- src/main/java/roomescape/reservation/service/PaymentService.java
@@ -1,39 +1,0 @@
-package roomescape.reservation.service;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.nio.charset.StandardCharsets;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestClient;
-import roomescape.common.exception.PaymentException;
-import roomescape.reservation.dto.request.TossPaymentConfirmRequest;
-import roomescape.reservation.dto.response.TossErrorResponse;
-import roomescape.reservation.dto.response.TossPaymentResponse;
-
-@Service
-public class PaymentService {
-    private final RestClient tossRestClient;
-    private final ObjectMapper objectMapper;
-
-    public PaymentService(RestClient tossRestClient, ObjectMapper objectMapper) {
-        this.tossRestClient = tossRestClient;
-        this.objectMapper = objectMapper;
-    }
-
-    public TossPaymentResponse confirmPayment(TossPaymentConfirmRequest request) {
-        return tossRestClient.post()
-                .uri("/payments/confirm")
-                .body(request)
-                .retrieve()
-                .onStatus(
-                        status -> status.is4xxClientError() || status.is5xxServerError(),
-                        (req, res) -> {
-                            String errorBody = new String(res.getBody().readAllBytes(), StandardCharsets.UTF_8);
-                            TossErrorResponse errorResponse = objectMapper.readValue(errorBody, TossErrorResponse.class);
-                            throw new PaymentException("결제 실패 : " + errorResponse.message());
-                        }
-                )
-                .body(TossPaymentResponse.class);
-    }
-
-
-}
